api.js moved in own directory
